Rename match details components
